% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllocateUnits.R
\docType{methods}
\name{AllocateUnits}
\alias{AllocateUnits}
\alias{AllocateUnits,UnitPrioritization,UnitAllocParam-method}
\title{Method to allocate units per population domain}
\usage{
AllocateUnits(object, Param)

\S4method{AllocateUnits}{UnitPrioritization,UnitAllocParam}(object, Param)
}
\arguments{
\item{object}{Object upon which the allocation is to be computed.}

\item{Param}{Parameters to allocate the units in each domain.}
}
\value{
Integer vector with the number of allocated units per domain.
}
\description{
Method to allocate units per population domain
}
\examples{
\dontrun{

UnitAllocParam <- new(Class = 'UnitAllocParam',
                      MaxAlloc = as.integer(c(23, 23, 45, 23, 12, 23, 26, 76, 12, 1, 1, 1)),
                      MinAlloc = rep(2L, length = 12),
                      AllocFactors = list(c(0.43, 0.12, 0.234, 0.123, 0.934, 0.23, 0.754, 0.423, 0.923, 0.345, 0.512, 0.623),
                                          c(0.43, 0.123, 0.34, 0.3, 0.34, 0.13, 0.954, 0.523, 0.223, 0.445, 0.522, 0.323)),
                      DomainWeights = c(0.34, 0.12, 0.1, 1, 123, 0.12, 56, 123, 98, 12, 23, 87),
                      MaxUnits = 265L )

AllocateUnits(PriorUnits, UnitAllocParam)
}
}
