% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookAllocMeasure.R
\docType{methods}
\name{CookAllocMeasure}
\alias{CookAllocMeasure}
\alias{CookAllocMeasure,contObsPredModelParam,CookAllocParam-method}
\title{Function to obtain the Cook-related allocation measure.}
\usage{
CookAllocMeasure(object, Param)

\S4method{CookAllocMeasure}{contObsPredModelParam,CookAllocParam}(object, Param)
}
\arguments{
\item{Variable}{numeric vector with the values of the variables}

\item{ObsPredParam}{object of class \linkS4class{ObsPredParam}}

\item{Factor}{data.table with the values of the variables working as factors}

\item{Prob}{numeric vector of length 1}

\item{alpha}{numeric vector of length 1}
}
\value{
It returns a numeric vector
}
\description{
Function to
}
\examples{
\dontrun{
ecfFun(c(1, 3, 5, 0, -1, 1))
}
}
