% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookAllocParam-class.R
\docType{class}
\name{CookAllocParam-class}
\alias{CookAllocParam-class}
\title{S4 class for the parameters of the Cook-related allocation proportional factor.}
\description{
Definition of the S4 class named \code{CookAllocParam} for the parameters for
the allocation of units to be edited interactively in each population domain under the
optimization approach to selective editing using the Cook distance in a linear regression model.
}
\section{Slots}{

\describe{
\item{\code{Domains}}{\linkS4class{data.table} with a row per domain with the values of the variables
determining the domains of the data set.}

\item{\code{MaxAlloc}}{Integer vector with the maximum number of units to be allocated in each domain.}

\item{\code{MinAlloc}}{Integer vector with the minimum number of units to be allocated in each domain.}

\item{\code{AllocFactors}}{List with numeric vectors as components containing the proportionality factor
of each domain for the allocation.}

\item{\code{MaxUnits}}{Integer vector of length 1 with the maximal number of units to allocate.}
}}
\examples{
CookAllocParam <- new(Class = 'CookAllocParam',
                      ObjVariables = c('CifraNeg_13.___', 'Personal_07._____705:1.-2.'),
                      Domains = 'ActivEcono_35._4._2.1.4._0',
                      Prob = 0.95,
                      alpha = 2 )

}

