% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AllocatedUnits.R
\docType{methods}
\name{plot,AllocatedUnits,character-method}
\alias{plot,AllocatedUnits,character-method}
\title{Plot a summary of the selection of units in the optimization approach to selective editing.}
\usage{
\S4method{plot}{AllocatedUnits,character}(x, y, ...)
}
\arguments{
\item{x}{Object of class \linkS4class{AllocatedUnits}.}

\item{y}{Charactec vector with the names of the variable names roots.}

\item{...}{Arguments to be passed to methods, especially including

\itemize{

 \item \code{Query} as a \linkS4class{data.table} with the values of either the identification
 variables of the queried statistical units or the domain variables of queried sample domain;

 \item \code{ModelParam} as a \linkS4class{data.table} with the parameters of the
 observation-prediction model;

 \item \code{UnitParam} as a \linkS4class{data.table} with the parameters of each statistical
 unit (sampling weights and unit scores).
}}
}
\value{
Returns an invisible \code{NULL} plotting the one scatterplot per variable.
}
\description{
This method overloads the plot function to draw a picture with the information about
the selection of influential statistical units in the optimization approach to selective editing.
}
\examples{

}
