% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllocatedUnits-class.R
\docType{class}
\name{AllocatedUnits-class}
\alias{AllocatedUnits-class}
\title{S4 class for the units allocated}
\description{
Definition of the S4 class named \code{AllocatedUnits} units allocated to be edited
interactively in each population domain under the optimization approach to selective editing.
}
\section{Slots}{

\describe{
\item{\code{Domains}}{\linkS4class{data.table} with a row per domain with the values of the variables
determining the domains of the data set.}

\item{\code{Units}}{list of \linkS4class{data.table}s with the statistical units per domain.}
}}
\examples{
new(Class = 'AllocatedUnits')

}

